Adds basic translator and api schemas (#27)
This migrates the translation logic from the PoC code in poc-showcasing
branch. Nothing special in here - just the manual translation plus the 
barebone API definitions for each OpenAI and AWS Bedrock.

Just in case anyway curious about why not using OpenAPI generator for
the OpenAI part, here's the comment in the apischema/openai package:
```
// Package openai contains the followings are the OpenAI API schema definitions.
// Note that we intentionally do not use the code generation tools like OpenAPI Generator not only to keep the code simple
// but also because the OpenAI's OpenAPI definition is not compliant with the spec and the existing tools do not work well.
```

Especially, the OpenAPI schema is a bit tricky especially on the
streaming=true/false
conditionals. Given that only a few endpoints will be supported here
anyways,
it's not that problematic to have the manually written data structure.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>