api: adds BackendRef into LLMBackendSpec (#40)
This adds `backendRef` to LLMBackendSpec which specifies
the "backend" either Service or Backend resource of Envoy Gateway.

The choice of not embedding is intentional - A backend can be a target
of routing in HTTPRoute and can be either of these two types. Hence
this is not suitable for embedding. In addition, in the implementation, 
we won't directly use or reference them, but just simply attach the 
necessary logic by the names, so basically no benefit by doing so.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>