api: do not embed HTTPRoute as-is (#47)
This is a follow up on #39, and stops embedding HTTPRoute as-is.
The rationale here is somewhat complicated; We have to know which
backend we route traffics to **before** Envoy decides the routing as
we need to do perform the transformation etc. Hence, we need to 
implement the routing logics (e.g. header matching, weights, etc.) 
by ourselves, and cannot rely on the native router code path in Envoy.
That results in us only being able to support only a small subset of 
HTTPRoute functionality as well as we cannot simply use the embedded
HTTPRoute as a "base" in the translation. If we embed HTTPRoute here, 
it would create an impression that  all HTTPRoute functionality can be 
supported  here no matter how much we document otherwise.

We discussed about this issue, and reached the consensus that we should
have our own LLMRoute rule definition and only have the fields that 
actually can be supported.

The example API would look like;
```
kind: LLMRoute
metadata:
  name: 
  namespace: default
spec:
  inputSchema:
    schema: OpenAI
  rules:
    - matches:
      - headers:
        - type: Exact
          name: x-envoy-ai-gateway-llm-model
          value: llama3-70b
      backendRefs:
        - name: kserve
          weight: 20
        - name: aws-bedrock
          weight: 80
```

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>