extproc: initial implementation (#50)
This scaffolds and implements the initial logic based on the 
current API. Notably, I intentionally organized the extproc code 
in a way that it is decoupled from EG concepts as well as the 
AI Gateway API level constructs. The rationale is that it not only
allows us to test the extproc without relying on not only EG/AIG
and even without kubernetes at all, but also some users are asking
for the use of outside the ai-gateway projects.

As for the implementation, there's nothing special going on, and 
it is simply mostly migrated from the MVP with the decoupling in mind.


I will follow up with the end-to-end test in subsequent PRs.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>