Scaffolds the initial extproc e2e test  (#58)
This scaffolds the initial extproc e2e tests using the secrets.
The GHA workflow is defined in a separate file so that we can run
the tests for PRs coming from fork repos as well under the condition
that
a maintainer verifies the code and adds the label "safe-to-test".

The actual test for AWS bedrock is commented out now as we will do the
request signing inside the extproc.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>