Implements AWS Request Signing inside extproc  (#61)
This commit implements the request signing logic in extproc
for AWS Bedrock support. Previously, in our PoC, we had used
Envoy's native AWS signing filter, but we faced a couple of issues,
so we decided to implement the signing in extproc which provides 
us much more flexibility.

The implementation here is pretty naive, and will be refactored 
after/during the implementation phase of the BackendSecurityPolicy
API #43. For example, it's currently just using env var credential
provider,
but this will change and become flexible soon.


This also updates the existing use of v1 AWS Go SDK to v2.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>