api: changes LLM prefix -> AI* (#76)
This comes up in a thread in #71 with @yuzisun, and we might
want to remove the LLM prefix. 

The rationale is that the current functionality is nothing to do 
with "LLM" but more about the general routing and authn/z
with "AI providers" where the input is OpenAI format.

On the other hand, there "will be" the LLM specific CRD such
as the configurations for prompt guard, semantics caching etc.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>