e2e: ensures the route recalculated (#135)
Previously, internally `x-envoy-ai-gateway-selected-backend`
was used for routing. However, it has some conflicts with the 
EG default setting named `suprressEnvoyHeaders: true` and 
that caused the route recalculation ignored. This fixes it.
In general, we shouldn't use the Envoy internal headers 
x-envoy-* for headers we generate in this project.

Thanks @yuzisun for finding this out.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>