translator: toolCalls support in OpenAI to Bedrock (#132)
In this PR, I improve the translator to handle toolCalls. If toolCalls
is not in the OpenAI message we append toolResultBlock for assistant
role message. In the Response body, we add toolCalls if bedrock content
has toolUse

I also refactor `openAIMessageToBedrockMessage` to have separate
functions to convert message for different roles `user`, `assistant`,
`system` and `tool`.

When dealing with image content in the roles message, add a function to
parse data uri and add proper error handling.

---------

Signed-off-by: yweng14 <yweng14@bloomberg.net>