extproc: allow configuring an UDS listen address (#149)
Allows configuring extproc to listen on an Unix domain socket. Note that
this PR just enables it in the extproc filter. Follow-up PRs will be
needed to accommodate that in the controller as well, but that would
require further changes to how the Gateway itself (not only the filter)
is deployed, to mount a common volume for the UDS.

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>