feat: Implementation of API Key Auth (#125)
This PR implements APIKey Auth for backendSecurityPolicy. The api key
stored in a secret will be mounted to the extproc's pod, and later
extracted by the appropriate AuthHandler.

Moved extprocDeployment into the sink as we need to tinker with the
deployment if backend/backendSecurityPolicy is updated.

Removed the envoy config yaml's field add_request_header and replaced it
with new API Key Auth implementation. CI passes which means that API Key
mounted to extproc + specified on backend auth works.

Will work on adding AWS Credential file after this.

---------

Signed-off-by: Aaron Choo <achoo30@bloomberg.net>
Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>
Co-authored-by: Takeshi Yoneda <t.y.mathetake@gmail.com>