translator: handling for OpenAI and AWSBedrock error response (#152)
Add handling for the OpenAI and Bedrock error response. Currently when
error response is returned it fails on decoding the `ConverseOutput` or
`OpenAIChatCompletionResponse` as it is assuming a 200 response.

For AWSBedrock the error response is converted to OpenAIError type when
returning to the user. For OpenAI error response we return directly back
without translation. In case of connection errors to the AWS Bedrock or
OpenAI backend, we translate to OpenAIError type based on the status
code and message body.

fixes https://github.com/envoyproxy/ai-gateway/issues/133

---------

Signed-off-by: Dan Sun <dsun20@bloomberg.net>