deps: upgrades go dependencies, ignore EG CVE (#174)
**Commit Message**:

This pull request includes updates to dependencies and 
configuration files for better security and compatibility. 
Envoy Gateway dependency is kept to the latest, whose 
tag is a weird v0.5.0-rc2, and because of that Trivy and
Dependabot tries to "downgrade" to the latest tagged 
version. This fixes that by adding EG to the ignore list of
dependency as well as add the latest EG CVE to the 
trivyignore file.

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>