translator: add toolUse support in streaming (#151)
In this PR, I convert `toolUse` in bedrock converse streaming response
to openai chat completions `tool_calls`. `ToolUseBlockDelta` is added
to`ConverseStreamEventContentBlockDelta` as documented in
[ContentBlockDelta](https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_ContentBlockDelta.html).
`ContentBlockStart` is added as documented in
[ContentBlockStart](https://docs.aws.amazon.com/bedrock/latest/APIReference/API_runtime_ContentBlockStart.html)

The role in `ChatCompletionResponseChunkChoiceDelta` is not nullable as
documented in
https://platform.openai.com/docs/api-reference/chat/streaming

Unit test is updated to cover streaming response with `toolUse`. The
base64 encoded streaming test data if from calling bedrock anthropic
Claude model with `toolConfig`.

---------

Signed-off-by: yweng14 <yweng14@bloomberg.net>
Co-authored-by: Dan Sun <dsun20@bloomberg.net>