test: testupstream returns openai fake responses (#198)
**Commit Message**:

As a part of documentation, we need a fake upstream server
that returns openai responses but without credentials.
This modifies the testupstream so that it can return the 
fake openai response when the response body is not given
by the client and the path is known.

To make the testupstream usable in the doctest, this will
start pushing the testupstream to GHCR.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>