docs: Moved Required tools section of Prerequisite to the top (#228)
**Commit Message**: This commit moves the required tools subsection to
the top of the "Prerequisite" page. Later in the page we ask users to
use kubectl and helm commands so we need to verify installations first.

<!--
Please write the commit message here, which
typically talks about the description / what this commit does / why we
need it

-->

**Related Issues/PRs (if applicable)**: n/a

<!--
Please add the related issues or PRs here.

n/a
-->

**Special notes for reviewers (if applicable)**: n/a

<!--
Please add any special notes for reviewers here.

n/a
-->


<!--
⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️
Please make sure that at least `make precommit test` passes before
submitting the PR as ready for review.
If there's anything you're unsure about, please ensure the PR is marked
as a draft. For example,
draft PRs would be ideal for discussing the approach to a problem or the
design of a feature.

Also, please make sure that you can check off all the items on the
checklist below.

Otherwise, you might unnecessarily consume the time of the maintainers
unless there's
a specific reason for not doing so.

**Checklist**:
- [ ] The PR title follows the same format as the merged PRs.
- [ ] I have read the [CONTRIBUTING.md](../CONTRIBUTING.md) (for
first-time contributors).
- [ ] I have made sure that `make precommit test` passes locally.
- [ ] I have written tests for any new line of code unless there's
existing tests that cover the changes.
- [ ] I have updated the documentation accordingly.
- [ ] I am sure the coding style is consistent with the rest of the
codebase.

⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️⚠️
-->

Signed-off-by: melsal13 <mmvsal13@gmail.com>