chore: reorganizes custom router/filterapi package (#234)
**Commit Message**:

This does the two re-organization of packages:
* Renames `filterconfig` to `filterapi` 
* Put `extprocapi` into `filterapi/x` package 

The `filterapi/x` package will serve as an experimental
package that facilitates the development of such features.
Notably, custom routers etc while keeping the ability to
introduce breaking changes at any release. This will be 
for advanced users as well as core contributors to iterate
on some advanced features in the single repo.

**Related Issues/PRs (if applicable)**:

For #233

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>