controller: refactors flags (#235)
**Commit Message**:

This refactors the main function of controller.
Previously it was conflated with the implementation detail
(zap) and that is exposed unnecessarily as controller main
arguments. This resolves it and expose the only necessary
one to have the consistency with extproc.

Notably, this also adds an new controller argument/helm 
value that configures the log level of external processor.
More fine tuned, per-route log level can be added later 
as in #204.

This also increases the test coverage.

**Related Issues/PRs (if applicable)**:

Supersedes #204

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>