extproc: decouple router package from api paths (#352)
**Commit Message**

extproc: decouple router package from api paths

This is a follow-up change to decouple the router package from the API
paths. Now that we have specialized processors per API path it does not
make sense for the router package to have path-related logic.
Now each processor is responsible for instantiating the right body
parser for the path they're processing.

**Related Issues/PRs (if applicable)**

Follow-up for https://github.com/envoyproxy/ai-gateway/pull/325 and
https://github.com/envoyproxy/ai-gateway/pull/334

**Special notes for reviewers (if applicable)**

N/A

---------

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>
Co-authored-by: Takeshi Yoneda <t.y.mathetake@gmail.com>