controller: removes sink implementation (#364)
**Commit Message**

This removes the config sink style implementation of controller.
Previously, there the config sink was used as a central place to do the
mutation that involves multiple resources. However, it unnecessarily
comes with the complexity and this simplifies the implementation by
removing it.

Majority of the changes are code moving, especially moving the resource
mutation code from sink to under AIGatewayRoute controller.

This was brought up in the conversation with @sanposhiho and more follow
up PRs will come to polish the implementation.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>