extproc: load default config if file does not exist (#376)
**Commit Message**

extproc: load default config if file does not exist

If the configuration file does not exist, let extproc run with the
default configuration. This prevents the extproc process from
terminating if the config does not exist, which is a valid use case when
the extproc filter is used in conjunction with other control planes.

**Related Issues/PRs (if applicable)**

N/A

**Special notes for reviewers (if applicable)**

N/A

---------

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>