controller: fixes extproc config update on removal (#380)
**Commit Message**

Previously, the deployment config is populated with the values specified
in the AIGatewayRoute resource and not removed if the values is not
specified. For example, after a replica value is specified, if a user
removes it from AIGatewayFilterConfig, then the replica is not reverted
to a default value. This fixes applyExtProcDeploymentConfigUpdate so
that it will remove the previously specified value.

Thanks for pointing out @sanposhiho !

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>