docs: removes --fail flags (#389)
**Commit Message**

At the first iteration of the "getting started guide", I added `--fail`
in all commands as we had an automated testing of the site documentation
and it allowed us to detect if the curl command not is working. However,
now we don't have any tests like that + `--fail` prevents users from
seeing the actual error response from the upstream which makes it
difficult to troubleshoot since it suppresses the response body due to
the early cancel of requests.

**Related Issues/PRs (if applicable)**

The root cause of #379 #375

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>