feat: store oidc as pointer for token providers (#415)
**Commit Message**

Set OIDC as a pointer in both OIDC Provider and
ClientCredentialsTokenProvider's field. Updates occurring to
OIDCProvider's oidcConfig will be "shared" with
ClientCredentialsTokenProvider (cctProvider).

**Issue**

oidcProvider and cctProvider share the same oidc values on init
(oidcProvider creates cctProvider and passes as param the same oidc
value). The issue occurs in `FetchToken` as `token_endpoint` of oidc
gets updated in oidcProvider and the subsequent call to cctProvider's
`FetchToken` expects `token_endpoint` to be propagated. The value of
`token_endpoint` in cctProvider is nil as no value is set to it.

Without propagating the value over, the error that shows up is: `failed
to get token: fail to get oauth2 token Post "token_endpoint":
unsupported protocol scheme ""`.

**Testing**
Updated the original FetchToken test to not set TokenEndpoint as part of
the oidc value. Original code results in the same error, but new changes
passed the test.

Signed-off-by: Aaron Choo <achoo30@bloomberg.net>